Labefy com musicas funcionando!

+++ modulo2/projeto-labefy/src/pages/ManageTracks/ManageTracks.js
@@ -2,7 +2,8 @@ import React from "react"
 import axios from "axios"
 import { BASE_URL } from "../../constants/urls"
 import { axiosConfig } from "../../constants/urls"
-import { TrackContainer, TitleContainer } from "./styled"
+import { TrackContainer, TitleContainer, SearchContainer, BodyContainer, Url } from "./styled"
+import Trashcan from "../../images/trashcan.svg"
 
 export default class ManageTracks extends React.Component {
     state = {
@@ -44,7 +45,7 @@ export default class ManageTracks extends React.Component {
             artist: this.state.trackArtistInput,
             url: this.state.trackUrlInput
         }
-        
+
         axios.post(`${BASE_URL}/${this.props.playlist.id}/tracks`, body, axiosConfig)
         .then((response) => {
             alert("Música adicionada com sucesso!")
@@ -57,6 +58,7 @@ export default class ManageTracks extends React.Component {
     }
 
     removeTrackFromPlaylist = (id) => {
+        if(window.confirm("Tem certeza que deseja remover essa música?"))
         axios.delete(`${BASE_URL}/${this.props.playlist.id}/tracks/${id}`, axiosConfig)
         .then((response) => {
             alert("Música deletada com sucesso!")
@@ -69,19 +71,28 @@ export default class ManageTracks extends React.Component {
 
     render () {
         const tracksList = this.state.tracks.map ((track) => {
+            let urlSplit = track.url.split("https://open.spotify.com/")
+            urlSplit[0] = "https://open.spotify.com/embed/"
+            const newUrl = urlSplit.join('')
+            
             return <TrackContainer key={track.id}>
+                <iframe 
+                    src={newUrl}
+                    allow="encrypted-media"
+                >
+                </iframe>
                 <p>{track.name}</p>
                 <p>{track.artist}</p>
-                <button onClick={() => this.removeTrackFromPlaylist(track.id)}>Remover Música</button>
+                <button onClick={() => this.removeTrackFromPlaylist(track.id)}><img src={Trashcan}/></button>
             </TrackContainer>
         })
 
-        return (<div>
+        return (<BodyContainer>
             <TitleContainer>
-                <h3>{this.props.playlist.name}</h3>
+                <h2>{this.props.playlist.name}</h2>
                 <button onClick={this.props.switchToPlaylists}>Voltar</button>
             </TitleContainer>
-            <div>
+            <SearchContainer>
                 <input
                     type="text"
                     value={this.state.trackNameInput}
@@ -94,17 +105,17 @@ export default class ManageTracks extends React.Component {
                     onChange={this.onChangeTrackArtist}
                     placeholder="Artista"
                 />
-                <input
+                <Url
                     type="url"
                     value={this.state.trackUrlInput}
                     onChange={this.onChangeTrackUrl}
                     placeholder="URL"
                 />
                 <button onClick={this.addTrackToPlaylist} >Adicionar Música</button>
-            </div>
+            </SearchContainer>
             <div>
                 {tracksList}
             </div>
-        </div>)
+        </BodyContainer>)
     }
 }
